+++ src/main/java/com/coderscampus/chatapp/a14/service/MessageService.java
@@ -23,7 +23,7 @@ public class MessageService {
 		return messageRepo.save(message);
 	}
 	
-	public List<Message> findMessagesForChannel(Long channelId) {
+	public List<Message> findMessagesForChannel(int channelId) {
 		Channel channel = channelRepo.findByChannelId(channelId);
 		return messageRepo.findMessagesByChannelId(channel);
 	}
